// src/app/users/models/create-user.model.ts
export class CreateUser {
  userName: string;
  password: string;
  commission: number;
  fullName: string;
  maindomain: string;
  domain: string;
  creditRef: number;
  userType: number;
  currencyCode: string;
  currencyId: number;
  sharePercent: number;
  prepaid:number;
  lcbd:number;
  lcbd1:number;
  allowDomainAllocation: number;
  allowRollingCommission: number;

  cricketSharing: number;
  cricketFancySharing: number;
  dogRaceSharing: number;
  horseRaceSharing: number;
  indianCasinoSharing: number;
  intCasinoSharing: number;
  soccerGoalsSharing: number;
  soccerSharing: number;
  tennisSharing: number;
  stakeSetting:string;
  captcha:number
}